added annotation and changed file names in remote/rmi
