+++ src/main/java/ru/ok/journal/remote/config/RemoteConfig.java
@@ -4,8 +4,8 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.remoting.rmi.RmiServiceExporter;
 import org.springframework.remoting.support.RemoteExporter;
+import ru.ok.journal.remote.rmi.IRMIComment;
 import ru.ok.journal.remote.rmi.RMIComment;
-import ru.ok.journal.remote.rmi.RMICommentImpl;
 
 @Configuration
 public class RemoteConfig {
@@ -18,8 +18,8 @@ public class RemoteConfig {
 
         RmiServiceExporter exporter = new RmiServiceExporter();
         exporter.setServiceName("PostCommentRMI");
-        exporter.setServiceInterface(RMIComment.class);
-        exporter.setService(new RMICommentImpl());
+        exporter.setServiceInterface(IRMIComment.class);
+        exporter.setService(new RMIComment());
 //        exporter.setService(new RMICommentImpl(id, postId, text));
 
         return exporter;
