+++ src/main/java/ru/ok/journal/remote/rmi/IRMIComment.java
@@ -1,0 +1,18 @@
+package ru.ok.journal.remote.rmi;
+
+
+public interface IRMIComment {
+
+    public long getId();
+
+    public void setId(long id);
+
+    public long getPostId();
+
+    public void setPostId(long postId);
+
+    public String getText();
+
+    public void setText(String text);
+
+}
