+++ src/main/java/ru/ok/journal/remote/rmi/RMIComment.java
@@ -1,17 +1,85 @@
 package ru.ok.journal.remote.rmi;
 
-public interface RMIComment {
+import org.springframework.stereotype.Service;
+import ru.ok.journal.model.Comment;
+import ru.ok.journal.model.Post;
+import ru.ok.journal.service.ICommentService;
+import ru.ok.journal.service.IPostService;
 
-    public long getId();
+import javax.annotation.Resource;
+import java.util.ArrayList;
+import java.util.List;
 
-    public void setId(long id);
+@Service
+public class RMIComment implements IRMIComment {
+    @Resource
+    ICommentService commentService;
+    @Resource
+    IPostService postService;
 
-    public long getPostId();
+    private long id;
+    private long postId;
+    private String text;
 
-    public void setPostId(long postId);
+    public RMIComment(long id, long postId, String text) {
+        this.id = id;
+        this.postId = postId;
+        this.text = text;
+    }
 
-    public String getText();
+    public RMIComment() {
+    }
 
-    public void setText(String text);
+    @Override
+    public long getId() {
+        return id;
+    }
+
+    @Override
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    @Override
+    public long getPostId() {
+        return postId;
+    }
+
+    @Override
+    public void setPostId(long postId) {
+        this.postId = postId;
+    }
+
+    @Override
+    public String getText() {
+        return text;
+    }
+
+    @Override
+    public void setText(String text) {
+        this.text = text;
+    }
+
+    public List<Comment> getByPost(List<Pair> postIds) {
+        //Pair - postID + lastCommentID
+        List<Post> posts = new ArrayList<>();
+        List<Comment> comments = new ArrayList<>();
+        List<Comment> tempComments = new ArrayList<>();
+        posts.addAll(postService.getAll());
+        for (Post post : posts) {
+            for (Pair postId : postIds) {
+                if (post.getId() == postId.getPostId()){
+                    tempComments.addAll(commentService.getByPost(post));
+                    for (Comment tempComment : tempComments) {
+                        if (tempComment.getId() >=postId.getLastCommentId()){
+                            comments.add(tempComment);
+                        }
+                    }
+                }
+            }
+        }
+
+        return comments;
+    }
 
 }
